Move pointer scanner items to its own folder
