+++ src/CelSerEngine.WpfBlazor/Components/PointerScanner/ModalPointerScanOptions.razor
