+++ src/CelSerEngine.WpfBlazor/Components/PointerScanner/PointerScanner.razor.js
